+++ internal/pkg/util/fs/overlay/overlay_item_linux.go
@@ -47,8 +47,13 @@ type Item struct {
 func NewItemFromString(overlayString string) (*Item, error) {
 	item := Item{Writable: true}
 
+	var err error
 	splitted := strings.SplitN(overlayString, ":", 2)
-	item.SourcePath = splitted[0]
+	item.SourcePath, err = filepath.Abs(splitted[0])
+	if err != nil {
+		return nil, fmt.Errorf("error while trying to convert overlay path %q to absolute path: %w", splitted[0], err)
+	}
+
 	if len(splitted) > 1 {
 		if splitted[1] == "ro" {
 			item.Writable = false
@@ -66,7 +71,7 @@ func NewItemFromString(overlayString string) (*Item, error) {
 	if s.IsDir() {
 		item.Type = image.SANDBOX
 	} else if err := item.analyzeImageFile(); err != nil {
-		return nil, fmt.Errorf("error encountered while examining image file %s: %s", item.SourcePath, err)
+		return nil, fmt.Errorf("while examining image file %s: %w", item.SourcePath, err)
 	}
 
 	return &item, nil
@@ -155,12 +160,12 @@ func (i *Item) mountDir() error {
 	}
 
 	if err = EnsureOverlayDir(i.StagingDir, false, 0); err != nil {
-		return fmt.Errorf("error accessing directory %s: %s", i.StagingDir, err)
+		return fmt.Errorf("error accessing directory %s: %w", i.StagingDir, err)
 	}
 
 	sylog.Debugf("Performing identity bind-mount of %q", i.StagingDir)
 	if err = syscall.Mount(i.StagingDir, i.StagingDir, "", syscall.MS_BIND, ""); err != nil {
-		return fmt.Errorf("failed to bind %s: %s", i.StagingDir, err)
+		return fmt.Errorf("failed to bind %s: %w", i.StagingDir, err)
 	}
 
 	// Best effort to cleanup mount
@@ -174,7 +179,7 @@ func (i *Item) mountDir() error {
 	// Try to perform remount
 	sylog.Debugf("Performing remount of %q", i.StagingDir)
 	if err = syscall.Mount("", i.StagingDir, "", syscall.MS_REMOUNT|syscall.MS_BIND, ""); err != nil {
-		return fmt.Errorf("failed to remount %s: %s", i.StagingDir, err)
+		return fmt.Errorf("failed to remount %s: %w", i.StagingDir, err)
 	}
 
 	return nil
@@ -187,14 +192,14 @@ func (i *Item) mountWithFuse(fuseMountTool string, additionalArgs ...string) err
 	var err error
 	fuseMountCmd, err := bin.FindBin(fuseMountTool)
 	if err != nil {
-		return fmt.Errorf("use of image %q as overlay requires %s to be installed: %s", i.SourcePath, fuseMountTool, err)
+		return fmt.Errorf("use of image %q as overlay requires %s to be installed: %w", i.SourcePath, fuseMountTool, err)
 	}
 
 	// Even though fusermount is not needed for this step, we shouldn't perform
 	// the mount unless we have the necessary tools to eventually unmount it
 	_, err = bin.FindBin("fusermount")
 	if err != nil {
-		return fmt.Errorf("use of image %q as overlay requires fusermount to be installed: %s", i.SourcePath, err)
+		return fmt.Errorf("use of image %q as overlay requires fusermount to be installed: %w", i.SourcePath, err)
 	}
 
 	// Obtain parent directory in which to create overlay-related mount
@@ -202,11 +207,11 @@ func (i *Item) mountWithFuse(fuseMountTool string, additionalArgs ...string) err
 	// related discussion.
 	parentDir, err := i.GetParentDir()
 	if err != nil {
-		return fmt.Errorf("error while trying to create parent dir for overlay %q: %s", i.SourcePath, err)
+		return fmt.Errorf("error while trying to create parent dir for overlay %q: %w", i.SourcePath, err)
 	}
 	fuseMountDir, err := os.MkdirTemp(parentDir, "overlay-mountpoint-")
 	if err != nil {
-		return fmt.Errorf("failed to create temporary dir %q for overlay %q: %s", fuseMountDir, i.SourcePath, err)
+		return fmt.Errorf("failed to create temporary dir %q for overlay %q: %w", fuseMountDir, i.SourcePath, err)
 	}
 
 	// Best effort to cleanup temporary dir
@@ -226,7 +231,7 @@ func (i *Item) mountWithFuse(fuseMountTool string, additionalArgs ...string) err
 	execCmd.Stderr = os.Stderr
 	_, err = execCmd.Output()
 	if err != nil {
-		return fmt.Errorf("encountered error while trying to mount image %q as overlay at %s: %s", i.SourcePath, fuseMountDir, err)
+		return fmt.Errorf("encountered error while trying to mount image %q as overlay at %s: %w", i.SourcePath, fuseMountDir, err)
 	}
 	i.StagingDir = fuseMountDir
 
@@ -259,18 +264,9 @@ func (i Item) unmountDir() error {
 // unmountFuse unmounts FUSE-based Items.
 func (i Item) unmountFuse() error {
 	defer os.Remove(i.StagingDir)
-	fusermountCmd, innerErr := bin.FindBin("fusermount")
-	if innerErr != nil {
-		// The code in performIndividualMounts() should not have created
-		// a FUSE-based overlay without fusermount in place
-		return fmt.Errorf("internal error: FUSE-based mount created without fusermount installed: %s", innerErr)
-	}
-	sylog.Debugf("Executing FUSE unmount command: %s -u %s", fusermountCmd, i.StagingDir)
-	execCmd := exec.Command(fusermountCmd, "-u", i.StagingDir)
-	execCmd.Stderr = os.Stderr
-	_, innerErr = execCmd.Output()
-	if innerErr != nil {
-		return fmt.Errorf("error while trying to unmount image %q from %s: %s", i.SourcePath, i.StagingDir, innerErr)
+	err := UnmountWithFuse(i.StagingDir)
+	if err != nil {
+		return fmt.Errorf("error while trying to unmount image %q from %s: %w", i.SourcePath, i.StagingDir, err)
 	}
 	return nil
 }